FIAKER: Fix Linux build errors and warnings

+++ modules/FiAKEr/iAFiberCharData.h
@@ -138,7 +138,7 @@ public:
 	QVector<double> maxAvgDifference;
 
 	//! IDs of the computed dissimilarity measures (ref. to position in getAvailableDissimilarityMeasures / switch in getDissimilarity
-	QVector<size_t> m_measures;
+	QVector<qulonglong> m_measures;
 
 	//! mapping for the resultID and projection error column:
 	uint m_resultIDColumn, m_projectionErrorColumn;
